requirements.txt
fastapi
sqlalchemy
asyncpg
databases
pydantic
uvicorn
pytest
httpx
pytest-asyncio
python-dotenv